+++ packages/single-line/src/SingleLineEditor.tsx
@@ -36,6 +36,16 @@ export interface SingleLineEditorProps {
    * sdk.locales
    */
   locales: LocalesAPI;
+
+  /**
+   * blur event handler
+   */
+  onBlur?: (event: React.FocusEvent<HTMLInputElement>) => void;
+
+  /**
+   * focus event handler
+   */
+  onFocus?: (event: React.FocusEvent<HTMLInputElement>) => void;
 }
 
 function isSupportedFieldTypes(val: string): val is 'Symbol' | 'Text' {
@@ -68,6 +78,8 @@ export function SingleLineEditor(props: SingleLineEditorProps) {
               isInvalid={errors.length > 0}
               isDisabled={disabled}
               value={value || ''}
+              onFocus={props.onFocus}
+              onBlur={props.onBlur}
               onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
                 setValue(e.target.value);
               }}