+++ packages/dropdown/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.14](https://github.com/contentful/field-editors/compare/@contentful/field-editor-dropdown@1.4.13...@contentful/field-editor-dropdown@1.4.14) (2024-06-13)
+
+**Note:** Version bump only for package @contentful/field-editor-dropdown
+
 ## [1.4.13](https://github.com/contentful/field-editors/compare/@contentful/field-editor-dropdown@1.4.12...@contentful/field-editor-dropdown@1.4.13) (2024-06-07)
 
 **Note:** Version bump only for package @contentful/field-editor-dropdown